Implement makefile
This Makefile takes a command line argument representing the strategy to
be used when reordering loops - in the future it will be called
multiple times by a wrapper script. The makefile builds an executable
from the unmodified files, as well as a reordered version.

Also includes example c files that can be used to show the strategies
and makefile in action.